DataTrustFramework/requirements.txt
pyspark>=3.5.0
delta-spark>=3.1.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
faker>=20.0.0
psutil>=5.0.0
kafka-python>=2.0.0
